Refactor auth module structure and implement SocialAuthService
