+++ backend/src/auth/auth.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { AuthService } from './auth.service';
-import { AuthController } from './auth.controller';
+import { AuthService } from './services/auth.service';
+import { AuthController } from './controllers/auth.controller';
 import { UserModule } from 'src/user/user.module';
 import { BcryptService } from './services/bcrypt.service';
 import { CryptoService } from './services/crypto.service';
@@ -9,10 +9,10 @@ import { JwtModule } from '@nestjs/jwt';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { SmsModule } from 'src/sms/sms.module';
 import { EmailModule } from 'src/email/email.module';
-import { SocialAuthController } from './social-auth.controller';
+import { SocialAuthController } from './controllers/social-auth.controller';
 import { GoogleStrategy } from './strategies/google.strategy';
 import { JwtAuthGuard } from './guards/auth.guard';
-import { SocialAuthService } from './social-auth.service';
+import { SocialAuthService } from './services/social-auth.service';
 
 @Module({
   controllers: [AuthController, SocialAuthController],
@@ -32,4 +32,4 @@ import { SocialAuthService } from './social-auth.service';
     }),
   ],
 })
-export class AuthModule {}
+export class AuthModule { }
